separate routes and move functions to modules
